+++ src/pint/models/parameter.py
@@ -507,6 +507,8 @@ class Parameter:
             log.warning(
                 f"Changing KOM from DT92 convention to IAU: this will not be readable by PINT"
             )
+        elif self.name == "DMDATA" and not format.lower() == "pint":
+            line = "%-15s %d" % (self.name, int(self.value))
 
         if self.uncertainty is not None:
             line += " %d %s" % (